:ok_hand: refactor: update jsx test file names
