+++ client/src/components/hoc/__tests__/authentication.test.jsx
