+++ client/src/components/hoc/authentication.jsx
@@ -1,63 +1,59 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Redirect } from 'react-router-dom';
 import PreLoader from '../shared/PreLoader';
 
 /**
- * @class AuthenticatedComponent
+ * @function AuthenticatedComponent
  * @return {Component} - MyComponent
+ * @param {object} props
  */
-class AuthenticatedComponent extends Component {
-  static propTypes = {
-    isAuthenticated: PropTypes.bool.isRequired,
-    location: PropTypes.shape({
-      hash: PropTypes.string,
-      key: PropTypes.string,
-      pathname: PropTypes.string,
-      search: PropTypes.string,
-      state: PropTypes.shape({
-        from: PropTypes.shape({
-          hash: PropTypes.string,
-          key: PropTypes.string,
-          pathname: PropTypes.string,
-          search: PropTypes.string
-        })
-      })
-    }).isRequired,
-    authLoading: PropTypes.bool.isRequired,
-    MyComponent: PropTypes.func.isRequired
-  };
+const AuthenticatedComponent = (props) => {
+  const {
+    MyComponent,
+    authLoading,
+    isAuthenticated,
+    location
+  } = props;
 
-  /**
-   * @return {Component} - MyComponent
-   */
-  render() {
-    const {
-      MyComponent,
-      authLoading,
-      isAuthenticated,
-      location
-    } = this.props;
-
-    if (authLoading) {
-      return <PreLoader />;
-    }
-
-    if (isAuthenticated) {
-      return <MyComponent {...this.props} />;
-    }
-
-    return (
-      <Redirect
-        to={{
-          pathname: '/login',
-          state: { from: location }
-        }}
-      />
-    );
+  if (authLoading) {
+    return <PreLoader />;
   }
-}
+
+  if (isAuthenticated) {
+    return <MyComponent {...props} />;
+  }
+
+  return (
+    <Redirect
+      to={{
+        pathname: '/login',
+        state: { from: location }
+      }}
+    />
+  );
+};
+
+AuthenticatedComponent.propTypes = {
+  isAuthenticated: PropTypes.bool.isRequired,
+  location: PropTypes.shape({
+    hash: PropTypes.string,
+    key: PropTypes.string,
+    pathname: PropTypes.string,
+    search: PropTypes.string,
+    state: PropTypes.shape({
+      from: PropTypes.shape({
+        hash: PropTypes.string,
+        key: PropTypes.string,
+        pathname: PropTypes.string,
+        search: PropTypes.string
+      })
+    })
+  }).isRequired,
+  authLoading: PropTypes.bool.isRequired,
+  MyComponent: PropTypes.func.isRequired
+};
 
 const requireAuthentication = (MyComponent) => {
   const mapStateToProps = state => ({
