+++ client/src/components/shared/Footer/__tests__/Footer.test.jsx
