+++ client/src/components/shared/StickyBar/__tests__/StickyBar.test.jsx
